==> 02-Assignment- Testing regular perceptron for XOR Gate.ipynb <==
import numpy as np

features = np.array([
    [0,0],
    [0,1],
    [1,0],
    [1,1]
])

#labels are results of OR gate
labels = np.array([0,1,1,0])

#initial weights can be anything
w = [0.9, 0.9]

threshold = 0.5

learning_rate = 1
epoch = 20 #learning time

for j in range(0, epoch):
    print("epoch: ",j)
    global_delta = 0
    for i in range(0, features.shape[0]):

        actual = labels[i]
        instance = features[i]

        x0 = instance[0]
        x1 = instance[1]

        sum_unit = x0 * w[0] + x1 * w[1]

        if sum_unit > threshold:
            fire = 1
        else:
            fire = 0

        delta = actual - fire
        global_delta = global_delta + abs(delta)

        print("prediction: ", fire, "whereas actual was ",actual, "(error:" , delta,")")

        w[0] = w[0] + delta * learning_rate
        w[1] = w[1] + delta * learning_rate
        
        
    print("------------------------------")
    
    if global_delta == 0:
        break

Tried with varying learning rates, tried with varying weights and changes number of epochs but it appears simple perceptron cannot solve XOR gate

from IPython.display import Image
Image(url= "perceptron_1.png", width=800, height=800)

As we can see, it calculates a weighted sum of its inputs and thresholds it with a step function. Geometrically, this means the perceptron can separate its input space with a hyperplane. That’s where the notion that a perceptron can only separate linearly separable problems came from. Since the XOR function is not linearly separable, it really is impossible for a single hyperplane to separate it.

from IPython.display import Image
Image(url= "XORgate.png", width=800, height=800)

This is why a regular perceptron cannot solve XOR gate. Also, an activation function instead of weighted sum to be applied in combination with backpropagation to be able to solve it
