==> 06 - pretrained models.ipynb <==
#### ResNet

pip install keras.applications

dir(keras.applications.resnet)

import numpy as np
import keras
from keras.preprocessing import image
from keras.applications import *



# Load Keras' ResNet50 model that was pre-trained against the ImageNet database
model = resnet.ResNet50()



# Load the image file, resizing it to 224x224 pixels (required by this model)
img = image.load_img("bay.jpg", target_size=(224, 224))



# Convert the image to a numpy array
x = image.img_to_array(img)



# Add a forth dimension since Keras expects a list of images
x = np.expand_dims(x, axis=0)


from keras.applications.resnet import preprocess_input


# Scale the input image to the range used in the trained network
x = preprocess_input(x)


# Run the image through the deep neural network to make a prediction
predictions = model.predict(x)


from keras.applications.resnet import decode_predictions

# Look up the names of the predicted classes. Index zero is the results for the first image.
predicted_classes = decode_predictions(predictions, top=9)




print("This is an image of:")

for imagenet_id, name, likelihood in predicted_classes[0]:
    print(" - {}: {:2f} likelihood".format(name, likelihood))
